+++ test/test_container.rb
@@ -5,7 +5,7 @@ require "pathname"
 require_relative "../lib/xnix/container"
 
 PROJECT_ROOT = Pathname.new(__dir__).join("..").realpath.to_s
-VERSION = "0.2.60"
+VERSION = "0.2.61"
 
 def assert(condition, message)
   return if condition